Fix CMake configure warnings

+++ code/icarus/IcarusImplementation.cpp
@@ -358,7 +358,7 @@ void CIcarus::Precache(char* buffer, long length)
 			{
 				float f = *(float *) block.GetMemberData( 0 );
 
-				if (f == TYPE_PATH)
+				if (f == (float)TYPE_PATH)
 				{
 					sVal1 = (const char *) block.GetMemberData( 1 );
