+++ code/rd-vanilla/tr_ghoul2.cpp
@@ -2301,7 +2301,7 @@ void RenderSurfaces(CRenderSurface &RS)
 //			&& RS.fogNum == 0
 			&& (RS.renderfx & RF_SHADOW_PLANE )
 			&& !(RS.renderfx & ( RF_NOSHADOW | RF_DEPTHHACK ) )
-			&& shader->sort == SS_OPAQUE )
+			&& shader->sort == (float)SS_OPAQUE )
 		{		// set the surface info to point at the where the transformed bone list is going to be for when the surface gets rendered out
 			CRenderableSurface *newSurf = AllocRS();
 			if (surface->numVerts >= SHADER_MAX_VERTEXES/2)
@@ -2322,7 +2322,7 @@ void RenderSurfaces(CRenderSurface &RS)
 //			&& RS.fogNum == 0
 			&& (RS.renderfx & RF_SHADOW_PLANE )
 			&& !(RS.renderfx & ( RF_NOSHADOW ) )
-			&& shader->sort == SS_OPAQUE )
+			&& shader->sort == (float)SS_OPAQUE )
 		{		// set the surface info to point at the where the transformed bone list is going to be for when the surface gets rendered out
 			CRenderableSurface *newSurf = AllocRS();
 			newSurf->surfaceData = surface;
